Tidying up logstic loss handling

+++ tests/models/test_logistic_regression_path.py
@@ -10,7 +10,7 @@ class TestLogisticRegressionPath(TestCase):
         X = np.array(
             [0.50, 0.75, 1.00, 1.25, 1.50, 1.75, 1.75, 2.00, 2.25, 2.50, 2.75, 3.00, 3.25, 3.50, 4.00, 4.25, 4.50, 4.75,
              5.00, 5.50])
-        y = np.array([0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1])
+        y = np.array([0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1], dtype=X.dtype)
         X = X[:, np.newaxis]
 
         self.X = X
@@ -35,7 +35,7 @@ class TestLogisticRegressionPath(TestCase):
     def test_pos_class(self):
         X = np.array(
             [0.50, 0.75, 1.00])
-        y = np.array([0, 1, 2])
+        y = np.array([0, 1, 2], dtype=X.dtype)
         X = X[:, np.newaxis]
 
         with self.assertRaises(ValueError):
